Adopt src-layout: add pyproject.toml, move modules to src/iMolpro, thin iMolpro.py launcher
