+++ build.ps1
@@ -13,6 +13,7 @@ if (-not(Test-Path -path $molpro_zip)) {
 
 
 cmd.exe /c conda install -c conda-forge -y --file=requirements.txt m2-base nsis
+pip install --no-build-isolation --no-deps .
 
 $versionfile = ( $env:TMP, "\VERSION") -join ""
 $PWD = (Get-Item .).FullName
